chore: 폴더 정리
- 다른 곳에서 참조하지 않는 스타일 중, 길이가 짧은 것들은 컴포넌트 파일과 합침
- 스타일과 컴포넌트만 있는 폴더는 파일 하나로 합침
