+++ client/src/common/LoadingFallback.tsx
@@ -1,8 +1,24 @@
+/** @jsxImportSource @emotion/react */
+
 import { css } from '@emotion/react';
+
 import { COLORS } from 'styles/colors';
 import { FONT_SIZE } from 'styles/sizes';
 
-export const loadingFallbackStyle = css({
+interface Props {
+  text: string;
+}
+
+export const LoadingFallback = ({ text }: Props) => {
+  return (
+    <div css={loadingFallbackStyle}>
+      <img src='/earlybird.gif' alt='loading earlybird' width={100} height={100} />
+      <span css={loadingTextStyle}>{text}</span>
+    </div>
+  );
+};
+
+const loadingFallbackStyle = css({
   width: '100%',
   height: '100%',
   display: 'flex',
@@ -16,7 +32,7 @@ export const loadingFallbackStyle = css({
   },
 });
 
-export const loadingTextStyle = css({
+const loadingTextStyle = css({
   marginTop: 20,
   fontWeight: 600,
   color: COLORS.TEXT_1,
