+++ client/src/common/Tooltip.tsx
@@ -1,9 +1,25 @@
+/** @jsxImportSource @emotion/react */
+
 import { css } from '@emotion/react';
 
 import { COLORS } from 'styles/colors';
 import { COMMON_SIZE } from 'styles/sizes';
 
-export const tooltipWrapperStyle = css({
+interface Props {
+  text: string;
+}
+
+export const Tooltip = ({ text }: Props) => {
+  return (
+    <div css={tooltipWrapperStyle}>
+      <div css={tooltipInnerStyle}>
+        <span>{text}</span>
+      </div>
+    </div>
+  );
+};
+
+const tooltipWrapperStyle = css({
   position: 'absolute',
   width: 'fit-content',
   filter: `drop-shadow(0 0 3px ${COLORS.SHADOW})`,
@@ -13,7 +29,7 @@ export const tooltipWrapperStyle = css({
   top: 32,
 });
 
-export const tooltipInnerStyle = css({
+const tooltipInnerStyle = css({
   width: 'fit-content',
   whiteSpace: 'nowrap',
   backgroundColor: COLORS.WHITE,
