--- client/src/common/Tooltip/index.tsx
@@ -1,17 +1,0 @@
-/** @jsxImportSource @emotion/react */
-
-import { tooltipInnerStyle, tooltipWrapperStyle } from './styles';
-
-interface Props {
-  text: string;
-}
-
-export const Tooltip = ({ text }: Props) => {
-  return (
-    <div css={tooltipWrapperStyle}>
-      <div css={tooltipInnerStyle}>
-        <span>{text}</span>
-      </div>
-    </div>
-  );
-};
