+++ client/src/pages/DetailPage/ViewModeContainer/InterestBadge.styles.ts
