--- client/src/pages/LoginPage/styles.ts
@@ -1,11 +1,0 @@
-import { css } from '@emotion/react';
-
-import { COLORS } from 'styles/colors';
-import { FONT_SIZE } from 'styles/sizes';
-
-export const loginPageSubHeaderStyle = css({
-  fontSize: FONT_SIZE.MEDIUM,
-  color: COLORS.TEXT_1,
-  textAlign: 'center',
-  marginBottom: 10,
-});
