--- client/src/pages/MainPage/mockData.ts
@@ -1,140 +1,0 @@
-export const mockData = [
-  {
-    id: '1',
-    language: 'JavaScript',
-    code: `export const DetailPage = () => {
-  const { id } = useParams();
-    return (
-    <div>
-      <MiniNavBar>
-        <>
-          <span>{id}</span>
-          <button type='button'>
-            <PencilIcon />
-          </button>
-        </>
-      </MiniNavBar>
-      상세페이지임 암튼그럼ㅋㅋ
-    </div>
-  );
-};`,
-    techStack: ['React', 'Emotion', 'Typescript'],
-    requirements: ['잠실사는사람만', '소통좋아해요'],
-    liked: false,
-  },
-  {
-    id: '2',
-    language: 'JavaScript',
-    code: `export const DetailPage = () => {
-  const { id } = useParams();
-    return (
-    <div>
-      <MiniNavBar>
-        <>
-          <span>{id}</span>
-          <button type='button'>
-            <PencilIcon />
-          </button>
-        </>
-      </MiniNavBar>
-      상세페이지임 암튼그럼ㅋㅋ
-    </div>
-  );
-};`,
-    techStack: ['React', 'Emotion', 'Typescript'],
-    requirements: ['잠실사는사람만', '소통좋아해요'],
-    liked: false,
-  },
-  {
-    id: '3',
-    language: 'JavaScript',
-    code: `export const DetailPage = () => {
-  const { id } = useParams();
-    return (
-    <div>
-      <MiniNavBar>
-        <>
-          <span>{id}</span>
-          <button type='button'>
-            <PencilIcon />
-          </button>
-        </>
-      </MiniNavBar>
-      상세페이지임 암튼그럼ㅋㅋ
-    </div>
-  );
-};`,
-    techStack: ['React', 'Emotion', 'Typescript'],
-    requirements: ['잠실사는사람만', '소통좋아해요'],
-    liked: false,
-  },
-  {
-    id: '4',
-    language: 'JavaScript',
-    code: `export const DetailPage = () => {
-  const { id } = useParams();
-    return (
-    <div>
-      <MiniNavBar>
-        <>
-          <span>{id}</span>
-          <button type='button'>
-            <PencilIcon />
-          </button>
-        </>
-      </MiniNavBar>
-      상세페이지임 암튼그럼ㅋㅋ
-    </div>
-  );
-};`,
-    techStack: ['React', 'Emotion', 'Typescript'],
-    requirements: ['잠실사는사람만', '소통좋아해요'],
-    liked: false,
-  },
-  {
-    id: '5',
-    language: 'JavaScript',
-    code: `export const DetailPage = () => {
-  const { id } = useParams();
-    return (
-    <div>
-      <MiniNavBar>
-        <>
-          <span>{id}</span>
-          <button type='button'>
-            <PencilIcon />
-          </button>
-        </>
-      </MiniNavBar>
-      상세페이지임 암튼그럼ㅋㅋ
-    </div>
-  );
-};`,
-    techStack: ['React', 'Emotion', 'Typescript'],
-    requirements: ['잠실사는사람만', '소통좋아해요'],
-    liked: false,
-  },
-  {
-    id: '6',
-    language: 'JavaScript',
-    code: `export const DetailPage = () => {
-  const { id } = useParams();
-    return (
-    <div>
-      <MiniNavBar>
-        <>
-          <span>{id}</span>
-          <button type='button'>
-            <PencilIcon />
-          </button>
-        </>
-      </MiniNavBar>
-      상세페이지임 암튼그럼ㅋㅋ
-    </div>
-  );
-};`,
-    techStack: ['React', 'Emotion', 'Typescript'],
-    requirements: ['잠실사는사람만', '소통좋아해요'],
-    liked: false,
-  },
-];
