--- client/src/pages/MessagePage/mockData.ts
@@ -1,56 +1,0 @@
-import { MessageListType, SingleMessageType } from 'types/message';
-
-export const chatListData: MessageListType = {
-  messages: [
-    { with: 'mincho', lastCheckTime: '2022-12-08T08:38:05.617Z' },
-    {
-      with: 'jiychoi',
-      lastCheckTime: '2022-12-07T11:15:00.617Z',
-    },
-    {
-      with: 'hgseo',
-      lastCheckTime: '2022-12-07T16:18:59.113Z',
-    },
-    {
-      with: 'limited-hyeon',
-      lastCheckTime: '2022-12-06T00:09:12.694Z',
-    },
-  ],
-  lastPageConnectTime: '2022-12-08T15:15:11.555Z',
-};
-
-export const singleMessageTypeMincho: SingleMessageType[] = [
-  { from: 'mincho', content: '하이하이', time: '2022-12-08T05:11:12.123Z' },
-  { from: 'mincho', content: '바이바이', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'mincho', content: '굿굿', time: '2022-12-08T05:18:12.111Z' },
-];
-
-export const singleMessageTypeJiychoi: SingleMessageType[] = [
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:11:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-  { from: 'jiychoi', content: 'ㅁㄴㅇㅁㄴㅇㄴㅁㅇㄴㅁㅇㅁㄴㅇㅁㄴ', time: '2022-12-08T05:15:12.123Z' },
-];
-
-export const singleMessageTypeHgseo: SingleMessageType[] = [
-  {
-    from: 'hgseo',
-    content:
-      'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-    time: '2022-12-08T05:11:12.123Z',
-  },
-];
-
-export const singleMessageTypeLimited: SingleMessageType[] = [
-  { from: 'limited-hyeon', content: '🌚', time: '2022-12-08T05:11:12.123Z' },
-];
