--- client/src/pages/SettingsPage/MiniCodeBox.styles.ts
@@ -1,9 +1,0 @@
-import { COMMON_SIZE } from 'styles/sizes';
-
-export const miniCodeBoxStyle = {
-  width: 230,
-  height: 160,
-  borderRadius: COMMON_SIZE.BORDER_RADIUS,
-  margin: 0,
-  overflow: 'hidden',
-};
